+++ resources/views/bids/edit-bid.blade.php
@@ -94,9 +94,7 @@
         </div> <!-- container container-xl -->
 
 
-
-
-
+        @include(\Scm\PluginBid\Facades\ScmPluginBid::getBladeRoot().'::bids/shared/file-dialog')
 
     @endsection
 @endcomponent